Correction in Task1.mlx file
From: -
"Next, try controlling the DC motors. The battery included in the kit has a rated voltage of 11.1 V, although the true value may be higher or lower depending on the charge level. You can use a voltmeter to check the exact voltage or use 11.1 as an approximate value."

To: -
"Next, try controlling the DC motors. The voltage value provided to the motors via the boost converter from the battery can be up to 12V".
Vmax = 12; %max motor voltage (Volts)